Arch overhaul 3

+++ src/main/java/io/github/robak132/mcrgb_forge/colors/HSL.java
@@ -65,15 +65,15 @@ public final class HSL implements Color {
         return this;
     }
 
-    public float hue() {
+    public int hue() {
         return h;
     }
 
-    public float saturation() {
+    public int saturation() {
         return s;
     }
 
-    public float lightness() {
+    public int lightness() {
         return l;
     }
